requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0
scikit-learn>=1.1.0

# Database
psycopg2-binary>=2.9.0
redis>=4.3.0

# NLP and ML
spacy>=3.4.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
transformers>=4.21.0

# Text processing
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0

# Data processing
openpyxl>=3.0.0
xlrd>=2.0.0
boto3>=1.24.0

# Web and API (optional)
flask>=2.2.0
flask-cors>=3.0.0
requests>=2.28.0

# Vector databases (optional)
pinecone-client>=2.2.0

# Utilities
python-dotenv>=0.20.0
tqdm>=4.64.0
click>=8.1.0

# Development
pytest>=7.1.0
pytest-cov>=3.0.0
black>=22.6.0
flake8>=5.0.0